Allow specifying the network type in include
