+++ autoPyTorch/pipeline/components/preprocessing/image_preprocessing/normalise/base_normalizer.py
@@ -39,8 +39,4 @@ class BaseNormalizer(autoPyTorchImagePreprocessingComponent):
     def __str__(self) -> str:
         """ Allow a nice understanding of what components where used """
         string = self.__class__.__name__
-        info = vars(self)
-        # Remove unwanted info
-        info.pop('random_state', None)
-        string += " (" + str(info) + ")"
         return string
