+++ autoPyTorch/pipeline/components/setup/early_preprocessor/EarlyPreprocessing.py
@@ -61,8 +61,4 @@ class EarlyPreprocessing(autoPyTorchSetupComponent):
     def __str__(self) -> str:
         """ Allow a nice understanding of what components where used """
         string = self.__class__.__name__
-        info = vars(self)
-        # Remove unwanted info
-        info.pop('random_state', None)
-        string += " (" + str(info) + ")"
         return string
