+++ autoPyTorch/pipeline/components/setup/network_backbone/base_network_backbone.py
@@ -4,13 +4,15 @@ from typing import Any, Dict, Set, Tuple
 import torch
 from torch import nn
 
+from autoPyTorch.constants import CLASSIFICATION_TASKS, STRING_TO_TASK_TYPES
 from autoPyTorch.pipeline.components.base_component import BaseEstimator
+from autoPyTorch.pipeline.components.base_choice import autoPyTorchChoice
 from autoPyTorch.pipeline.components.base_component import (
     autoPyTorchComponent,
 )
 
 
-class BaseBackbone(autoPyTorchComponent):
+class NetworkBackboneComponent(autoPyTorchComponent):
     """
     Backbone base class
     """
@@ -26,7 +28,23 @@ class BaseBackbone(autoPyTorchComponent):
         """
         Not used. Just for API compatibility.
         """
+        input_shape = X['X_train'].shape[1:]
+
+        self.backbone = self.build_backbone(
+            input_shape=input_shape,
+        )
         return self
+
+    def transform(self, X: Dict[str, Any]) -> Dict[str, Any]:
+        """
+        Adds the scheduler into the fit dictionary 'X' and returns it.
+        Args:
+            X (Dict[str, Any]): 'X' dictionary
+        Returns:
+            (Dict[str, Any]): the updated 'X' dictionary
+        """
+        X.update({'network_backbone': self.backbone})
+        return X
 
     @abstractmethod
     def build_backbone(self, input_shape: Tuple[int, ...]) -> nn.Module:
