+++ autoPyTorch/pipeline/components/setup/network_head/fully_connected.py
@@ -8,7 +8,9 @@ import numpy as np
 
 from torch import nn
 
-from autoPyTorch.pipeline.components.setup.network.head.base_head import BaseHead
+from autoPyTorch.pipeline.components.setup.network_head.base_network_head import (
+    NetworkHeadComponent,
+)
 
 _activations: Dict[str, nn.Module] = {
     "relu": nn.ReLU,
@@ -17,7 +19,7 @@ _activations: Dict[str, nn.Module] = {
 }
 
 
-class FullyConnectedHead(BaseHead):
+class FullyConnectedHead(NetworkHeadComponent):
     """
     Standard head consisting of a number of fully connected layers.
     Flattens any input in a array of shape [B, prod(input_shape)].
@@ -44,6 +46,9 @@ class FullyConnectedHead(BaseHead):
         return {
             'shortname': 'FullyConnectedHead',
             'name': 'FullyConnectedHead',
+            'handles_tabular': True,
+            'handles_image': False,
+            'handles_time_series': False,
         }
 
     @staticmethod
