+++ autoPyTorch/pipeline/components/setup/network_initializer/base_network_initializer.py
@@ -91,9 +91,4 @@ class BaseNetworkInitializerComponent(autoPyTorchSetupComponent):
     def __str__(self) -> str:
         """ Allow a nice understanding of what components where used """
         string = self.__class__.__name__
-        info = vars(self)
-        # Remove unwanted info
-        info.pop('strategy', None)
-        info.pop('random_state', None)
-        string += " (" + str(info) + ")"
         return string
