+++ autoPyTorch/pipeline/components/setup/traditional_ml/tabular_classifier.py
@@ -63,9 +63,4 @@ class TabularClassifier(BaseModelComponent):
 
     def __str__(self) -> str:
         """ Allow a nice understanding of what components where used """
-        info = vars(self)
-        # Remove unwanted info
-        info.pop('random_state', None)
-        info.pop('fit_output', None)
-        info.pop('config', None)
-        return f"TabularClassifier: {self.model.name if self.model is not None else None} ({str(info)})"
+        return f"TabularClassifier: {self.model.name if self.model is not None else None}"
