+++ autoPyTorch/pipeline/components/training/trainer/base_trainer_choice.py
@@ -67,7 +67,6 @@ class TrainerChoice(autoPyTorchChoice):
         self._fit_requirements: Optional[List[FitRequirement]] = [
             FitRequirement("lr_scheduler", (_LRScheduler,), user_defined=False, dataset_property=False),
             FitRequirement("job_id", (str,), user_defined=False, dataset_property=False),
-            FitRequirement("network", (torch.nn.Sequential,), user_defined=False, dataset_property=False),
             FitRequirement(
                 "optimizer", (Optimizer,), user_defined=False, dataset_property=False),
             FitRequirement("train_data_loader",
